disp2: sunxi: update driver
1. verify smart backlight
2. verify enhance
3. verify capture(rgb/yv12/nv12)
4. add lcd pinctrl state manage

+++ drivers/video/sunxi/disp2/disp/de/lowlevel_sun50iw1/de_bws.c
@@ -20,7 +20,7 @@ static volatile __bws_reg_t *bws_dev[DEVICE_NUM][CHN_NUM];
 static de_reg_blocks bws_block[DEVICE_NUM][CHN_NUM];
 static de_reg_blocks bws_para_block[DEVICE_NUM][CHN_NUM];
 
-static unsigned int bws_hw_base[DEVICE_NUM][CHN_NUM] = {{0}};
+static uintptr_t bws_hw_base[DEVICE_NUM][CHN_NUM] = {{0}};
 
 //for bws
 __bws_status_t *g_bws_status[DEVICE_NUM][CHN_NUM];
